Split up the files a bit more, and cleaned up unused generics from ALCServer
