+++ corefxlab.sln
@@ -121,9 +121,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "System.Text.CaseFolding", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ALCProxy.TestAssembly", "tests\ALCProxy.TestAssembly\ALCProxy.TestAssembly.csproj", "{8F6F600E-5EFB-44B3-80DE-808974A2998A}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ALCProxy.TestInterface", "tests\ALCProxy.TestInterface\ALCProxy.TestInterface.csproj", "{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ALCProxy.TestInterfaceUpdated", "tests\ALCProxy.TestInterfaceUpdated\ALCProxy.TestInterfaceUpdated.csproj", "{14FDEB76-14AE-4332-819C-1E8054855469}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ALCProxy.TestInterfaceUpdated", "tests\ALCProxy.TestInterfaceUpdated\ALCProxy.TestInterfaceUpdated.csproj", "{14FDEB76-14AE-4332-819C-1E8054855469}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ALCProxy.TestInterface", "tests\ALCProxy.TestInterface\ALCProxy.TestInterface.csproj", "{6435C875-F526-45E1-97E1-B5BF8C4F9A55}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -771,18 +771,6 @@ Global
 		{8F6F600E-5EFB-44B3-80DE-808974A2998A}.Release|x64.Build.0 = Release|Any CPU
 		{8F6F600E-5EFB-44B3-80DE-808974A2998A}.Release|x86.ActiveCfg = Release|Any CPU
 		{8F6F600E-5EFB-44B3-80DE-808974A2998A}.Release|x86.Build.0 = Release|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Debug|x64.Build.0 = Debug|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Debug|x86.Build.0 = Debug|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Release|x64.ActiveCfg = Release|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Release|x64.Build.0 = Release|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Release|x86.ActiveCfg = Release|Any CPU
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081}.Release|x86.Build.0 = Release|Any CPU
 		{14FDEB76-14AE-4332-819C-1E8054855469}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{14FDEB76-14AE-4332-819C-1E8054855469}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{14FDEB76-14AE-4332-819C-1E8054855469}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -795,6 +783,18 @@ Global
 		{14FDEB76-14AE-4332-819C-1E8054855469}.Release|x64.Build.0 = Release|Any CPU
 		{14FDEB76-14AE-4332-819C-1E8054855469}.Release|x86.ActiveCfg = Release|Any CPU
 		{14FDEB76-14AE-4332-819C-1E8054855469}.Release|x86.Build.0 = Release|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Debug|x64.Build.0 = Debug|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Debug|x86.Build.0 = Debug|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Release|x64.ActiveCfg = Release|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Release|x64.Build.0 = Release|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Release|x86.ActiveCfg = Release|Any CPU
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -853,8 +853,8 @@ Global
 		{D76469F4-3694-463A-8444-187080ECD752} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
 		{5AD1F0B7-6279-45CE-BE5A-D35CCD198BAC} = {4B000021-5278-4F2A-B734-DE49F55D4024}
 		{8F6F600E-5EFB-44B3-80DE-808974A2998A} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
-		{9127FE29-3433-4EAD-AA6A-FA7BED3C1081} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
 		{14FDEB76-14AE-4332-819C-1E8054855469} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
+		{6435C875-F526-45E1-97E1-B5BF8C4F9A55} = {3079E458-D0E6-4F99-8CAB-80011D35C7DA}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9DD4022C-A010-4A9B-BCC5-171566D4CB17}
